Hooked up Honeycomb
In this commit, I completed the steps to have HoneyComb
capture data for the application. I added a tracing.js
file that initializes the HoneyComb SDK, and edited the
"start" command so that it runs the HoneyComb SDK. If we
need to test drive this while testing we can run
"yarn run build" then "yarn run start". "yarn run dev"
is what we usually run for development which interprets
the TS files.