Tweak Jest to ignore JS files and target only spec.ts files
By default, Jest was gladly picking up JS files from our output
folder and running those as tests. This wasn't what we wanted.
We only want to run tests against the Typescript code, not the
generated Javascript one. The generated Javascript isn't something
we look at or maintain, we're trusting Babel (previously TSC)
to spit out that code.

The second problem was that Jest was also running all files in
the __tests__ directories. We have factories in that directory
and don't want Jest treating those as tests, so I updated the
configuration to only target .spec.ts[x] files.