Added Babel to support using async/await in tests
I added Babel to allow use of async/await in tests
because Jest doesn't understand Typescript syntax.
It only understands Javascript syntax. As such, we
need something that can convert it to vanilla
Javascript so that Jest can understand it.

Babel is a great tool for achieving this. Babel
is a Javascript compiler in general. It supports
much of the newer syntax in ES and also facilitates
transpilation of Typescript to Javascript, which
is one of the things we're particularly interested
in here.

I installed Babel, and in the configuration, I'm using
two presets:
- babel-env - this is a somewhat magical/powerful preset
  that allows me to tell Babel that I'm targeting the
  current version of NodeJS. Do what you need to do transpile
  and polyfill things for that version of NodeJS.
- typescript - the Typescript preset enables the transpilation
  over Typescript code down to vanilla JS.

One question I'm unsure about is how I was able to use
async/await before I installed Babel. I was using ts-node
to interpret the Typescript. I'm uncertain whether async/await
is natively supported by ts-node, or there's something else at
play facilitating the transpilation. I'll need to look into that.

Jest doesn't support Typescript out of the box. If you
try to run