Refactored LLMSuggestionService to increase testability
The LLMSuggestionService previously used the LLMClient
directly without any mechanism for swapping it out with
a test double or fake. In this commit, I've refactored
the class to ensure that we can inject the LLMClient
and the PromptTemplate that the suggestion service
uses.

This will allow me to inject a configurable mock
for our tests, without having to reach inside the
LLMService and modify what it's using.

Additionally, I made the schema exportable so that
we could access it in the tests.