Renamed type for clarity - changed SchemaDefinition to ObjectForSchema
I was calling the type for valid objects for some schema, the schema
definition, which is incorrect. Schema definition is an alias for
schema. So effectively, my Zod type, was a schema that had a schema
definition. That doesn't make sense.

There's a schema and there's an object that matches the schema.