Renamed ambigious variable for clarity - structured output
Previously, I was calling the output from the LLMClient the
structured output. This isn't incorrect, but it is ambiguous
and doesn't really communicate much about what it contains.

Instead, I've adopted the term "RemoteGetSuggestionsResponse",
which better describes what it is. We've gotten some information
back from the LLMClient and that's the response we're looking
for.

I also added a type specifically for this. It's derived from
the schema definition.