Refactored tests, and enriched factories for remote responses
I refactored the tests a bit so that they're smaller and more
digestable. To facilitate this, I added a factory for the
generation of remote LLM Client responses. This removes a
couple lines from the test and makes it less verbose.

I also added a test to verify the behavior of the service
when there's an error.